+++ src/main/java/com/localstrategy/util/types/SingleTransaction.java
@@ -1,5 +1,7 @@
 package com.localstrategy.util.types;
 
+//TODO: How are walls detectable irl? Can we still stream and 'dYnaMyCallY' detect them as soon as they come?
+
 public record SingleTransaction(
         double price,
         double amount,